requirements.txt
backports.shutil-get-terminal-size==1.0.0
cffi==1.7.0
crypto==1.4.1
cryptography==1.4
cssselect==0.9.2
decorator==4.0.10
Django==1.9.12
django-debug-toolbar==1.4
django-extensions==1.6.7
django-template-debug==0.3.4
ecdsa==0.13
enum34==1.1.6
eutils==0.1.0
Fabric==1.11.1
idna==2.1
ipaddress==1.0.16
ipython==4.2.1
ipython-genutils==0.1.0
lxml==3.6.0
metapub==0.4.1
Naked==0.1.31
paramiko==1.17.1
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.2
pluggy==0.3.1
psycopg2==2.5.3
ptyprocess==0.5.1
py==1.4.31
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
python-memcached==1.53
pytz==2016.4
PyYAML==3.11
requests==2.10.0
shellescape==3.4.1
simplegeneric==0.8.1
six==1.10.0
South==0.8.4
sqlparse==0.1.19
tabulate==0.7.5
tox==2.3.1
traitlets==4.2.2
Unidecode==0.4.19
virtualenv==15.0.2
Werkzeug==0.11.10
wheel==0.24.0
django-celery==3.2.1
celery==4.0.2
django-dotenv==1.4.1
flower==0.9.1
redis==2.10.5
hg+https://bitbucket.org/fabian/django-taggit-autosuggest/#egg=django_taggit_autosuggest
biopython==1.67
Unipath==1.1